renamed GetStrike to StrikeSearch
